vfs: make O_PATH file descriptors usable for 'fstat()'
We already use them for openat() and friends, but fstat() also wants to
be able to use O_PATH file descriptors.  This should make it more
directly comparable to the O_SEARCH of Solaris.

Note that you could already do the same thing with "fstatat()" and an
empty path, but just doing "fstat()" directly is simpler and faster, so
there is no reason not to just allow it directly.

See also commit 332a2e1244bd, which did the same thing for fchdir, for
the same reasons.

Reported-by: ольга крыжановская <olga.kryzhanovska@gmail.com>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: stable@kernel.org    # O_PATH introduced in 3.0+
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
[haggertk: This is essentially a "reland". This change was originally
 backported in 3.4.12, but R really wants the fget_raw_light() call.
 It just so happens that the original version of this change
 (non-backport) that exists in 3.6+ has that proper call]
Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Change-Id: Ica68b3cef47401c0f973fa4bdc6b92581a4d1d7b
Signed-off-by: Sung Mingi <FiestaLake@protonmail.com>